+++ AdjusterMain.py
@@ -27,6 +27,7 @@ def adjust(args):
     apply_rom_settings(rom, args.heartbeep, args.heartcolor, args.quickswap, args.fastmenu, args.disablemusic,
                        args.sprite, args.ow_palettes, args.uw_palettes, args.reduce_flashing)
 
+    output_path.cached_path = args.outputpath
     rom.write_to_file(output_path('%s.sfc' % outfilebase))
 
     logger.info('Done. Enjoy.')