+++ source/gui/adjust/overview.py
@@ -111,6 +111,7 @@ def adjust_page(top, parent, settings):
         guiargs.rom = self.romVar2.get()
         guiargs.baserom = top.pages["randomizer"].pages["generation"].widgets["rom"].storageVar.get()
         guiargs.sprite = self.sprite
+        guiargs.outputpath = os.path.dirname(guiargs.rom)
         try:
             adjust(args=guiargs)
         except Exception as e: