+++ RELEASENOTES.md
@@ -17,6 +17,7 @@ Thanks to qadan, cheuer, & compiling
 * 0.4.0.7
 	* Reduce flashing option added
 	* Sprite author credit added
+	* Ranged Crystal switch rules tweaked
 * 0.4.0.6
 	* Hints now default to off
 	* The maiden gives you a hint to the attic if you bring her to the unlit boss room